attempt mass move of replay parser up a module level
