+++ metamon/data/replay_dataset/raw_replays/may_metamon_dataset.json
@@ -1,0 +1,106 @@
+{
+  "gen3ubers": [
+    {
+      "month": "2025-05",
+      "count": 66
+    }
+  ],
+  "gen4ou": [
+    {
+      "month": "2025-05",
+      "count": 1756
+    }
+  ],
+  "gen4nu": [
+    {
+      "month": "2025-05",
+      "count": 37
+    }
+  ],
+  "gen1nu": [
+    {
+      "month": "2025-05",
+      "count": 65
+    }
+  ],
+  "gen4ubers": [
+    {
+      "month": "2025-05",
+      "count": 311
+    }
+  ],
+  "gen3uu": [
+    {
+      "month": "2025-05",
+      "count": 27
+    }
+  ],
+  "gen2ou": [
+    {
+      "month": "2025-05",
+      "count": 394
+    }
+  ],
+  "gen2ubers": [
+    {
+      "month": "2025-05",
+      "count": 13
+    }
+  ],
+  "gen1uu": [
+    {
+      "month": "2025-05",
+      "count": 28
+    }
+  ],
+  "gen4uu": [
+    {
+      "month": "2025-05",
+      "count": 34
+    }
+  ],
+  "gen1ou": [
+    {
+      "month": "2025-05",
+      "count": 833
+    }
+  ],
+  "gen1ubers": [
+    {
+      "month": "2025-05",
+      "count": 43
+    }
+  ],
+  "gen3nu": [
+    {
+      "month": "2025-05",
+      "count": 43
+    }
+  ],
+  "gen3ou": [
+    {
+      "month": "2025-05",
+      "count": 4053
+    }
+  ],
+  "gen2uu": [
+    {
+      "month": "2025-05",
+      "count": 63
+    }
+  ],
+  "gen2nu": [
+    {
+      "month": "2025-05",
+      "count": 5
+    }
+  ],
+  "_metadata": {
+    "created_at": "2025-05-29T13:22:24.261365",
+    "total_formats": 15,
+    "date_range": {
+      "start": "2025-05",
+      "end": "2025-05"
+    }
+  }
+}
