+++ metamon/data/replay_dataset/replay_parser/checks.py
@@ -1,5 +1,5 @@
-from metamon.data.replay_dataset.parsed_replays.replay_parser.exceptions import *
-from metamon.data.replay_dataset.parsed_replays.replay_parser.forward import (
+from metamon.data.replay_dataset.replay_parser.exceptions import *
+from metamon.data.replay_dataset.replay_parser.forward import (
     SpecialCategories,
 )
 from poke_env.environment import Effect as PEEffect
